pyproject.toml
[project]
name = "pywjs-backend"
version = "0.1.1"
description = "The FastAPI Backend for the pywjs core."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioboto3>=14.3.0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "greenlet>=3.2.2",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.4",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "python-ulid[pydantic]>=3.0.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.9",
]
test = [
    "aiosqlite>=0.21.0",
    "anyio>=4.9.0",
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
]

[tool.ruff]
line-length = 88


[tool.pytest.ini_options]
addopts = [
    "--strict-config",
    "--strict-markers",
    "--ignore=docs_src",
]
